updates (#92)
* improve: onboarding

* feat: Add `project` endpoints

* add: accessReq table & implement usage of waitList

* fix: Fix creating/deleting `projects`

* prepare local electric db

* fix: Fix projects

* feat: add animation to members toolbar

* content placeholder in tables

* fix: DnD menu nav bugs

* improve: DnD menu nav

* improve: preselect the first option Add project invite users

* fix configs

* +

* +

* +

* +

* improve: use-hot-keys hook

* +

* +

* +

* +

* Implement initial version of "permission-manager" module, providing an abstraction layer for RBAC

* add: react-sticky-box & remove react-sticky-el

* improve: start fix projects && tasks loading

* update readme and lint fixes

* add focus on toggleGroup

* Refactor organizationTenantGuard by replacing it with setCtxt and protect middleware functions

* Refactor workspaceTenantGuard by replacing it with setCtxt and protect middleware functions

* Replace raw authentication with authGuard middleware

* Add TODOs to streamline resource identification and role hierarchy management

* Store user memberships in context for API query utilization

* Remove user as a potential context from ctx

* add: requestTable & change n8n endpoints to action-request(hono)

* Consolidate 'setCtx' logic into 'protect' middleware to streamline guard implementation

* fix: github auth bag

* clean up docker

* change readme further

* remove uuid

* uuid to nano

* allow app to function without local db

* move electric to own folder

* put electric db request in useEffect

* check for avail Electric everywhere

* dont rebuild app when connected

* fix: Fix migrations

* sort translations

* improve: if opening cella for first time redirect welcome instead of home

* improve: signout remove drafts in store too

* improve: welcome completed

* improve: organizationsTable role select order

* oauthOptions = oauthProviders

* feat: start transition to framer motion

* fix: remove weird animations

* add: requests Table

* improve: requests Table

* improve: board view filter

* fix: Fix electric and migration

* change: leaflet to @vis.gl/react-google-maps

* fix: Fix env example

* update deps

* improve: contact-form-map

* improve: env

* feat: finish migrations to framer motion

* some refactoring

* cleanup

* feat: add tabs animation

* fix: solve layout animation bug

* diable sticky on small screens

* fix: Fix tasks

* fix: Fix scripts

* fix: Remove generated and add supabase compose file

* add: slack notification

* improve: requests table

* Update README.md

* feat: Extend organizations seed

* Update README.md

* fix: Fix projects

* improve: requests table

* add: custom zoom contact-form-map

* improve: requests table

---------

+++ frontend/src/modules/projects/board-header.tsx
@@ -1,6 +1,7 @@
-import { PanelTopClose, Plus, Settings, Tag, Trash, XSquare, FilterX } from 'lucide-react';
+import { FilterX, PanelTopClose, Plus, Settings, Tag, Trash, XSquare } from 'lucide-react';
 import { useContext, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
+import { toast } from 'sonner';
 import { dialog } from '~/modules/common/dialoger/state';
 import { FocusView } from '~/modules/common/focus-view';
 import { sheet } from '~/modules/common/sheeter/state';
@@ -9,15 +10,14 @@ import DisplayOptions from '~/modules/projects/display-options';
 import WorkspaceView from '~/modules/projects/view-options';
 import { Button } from '~/modules/ui/button';
 import { WorkspaceSettings } from '~/modules/workspaces/workspace-settings';
+import { AvatarWrap } from '../common/avatar-wrap';
+import { FilterBarActions, FilterBarContent, TableFilterBar } from '../common/data-table/table-filter-bar';
+import { useElectric } from '../common/electric/electrify';
+import { TooltipButton } from '../common/tooltip-button';
 import { Badge } from '../ui/badge';
 import { WorkspaceContext } from '../workspaces';
 import AddProjects from './add-projects';
 import LabelsTable from './labels-table';
-import { type Label, useElectric } from '../common/root/electric';
-import { AvatarWrap } from '../common/avatar-wrap';
-import { FilterBarActions, FilterBarContent, TableFilterBar } from '../common/data-table/table-filter-bar';
-import { TooltipButton } from '../common/tooltip-button';
-import { toast } from 'sonner';
 
 interface BoardHeaderProps {
   showPageHeader: boolean;
@@ -27,12 +27,10 @@ interface BoardHeaderProps {
 const BoardHeader = ({ showPageHeader, handleShowPageHeader }: BoardHeaderProps) => {
   const { t } = useTranslation();
 
-  const { workspace, selectedTasks, setSelectedTasks, projects, searchQuery, tasks, setSearchQuery } = useContext(WorkspaceContext);
+  const { workspace, selectedTasks, setSelectedTasks, searchQuery, tasks, setSearchQuery, labels } = useContext(WorkspaceContext);
 
   // biome-ignore lint/style/noNonNullAssertion: <explanation>
-  const { db } = useElectric()!;
-
-  const labels = projects.flatMap((project) => project.labels).filter(Boolean) as Label[];
+  const Electric = useElectric()!;
 
   const openSettingsSheet = () => {
     sheet(<WorkspaceSettings sheet />, {
@@ -53,7 +51,9 @@ const BoardHeader = ({ showPageHeader, handleShowPageHeader }: BoardHeaderProps)
   };
 
   const onRemove = () => {
-    db.tasks
+    if (!Electric) return toast.error(t('common:no_local_db'))
+
+    Electric.db.tasks
       .deleteMany({
         where: {
           id: {
